requirements.txt
joblib==1.1.0
numpy==1.19.5
pandas==1.3.4
Pillow==9.0.0
plotly==5.4.0
streamlit==1.2.0
streamlit_player==0.1.5
